fix: rename enums file
