+++ external-import/dragos/dragos/domain/models/octi/enums.py
