+++ external-import/dragos/tests/tests_dragos/test_domain/test_models/test_domain.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 import pytest
 import stix2
 import dragos.domain.models.octi as octi
-import dragos.domain.models.octi.enum as octi_enums
+import dragos.domain.models.octi.enums as octi_enums
 from pydantic import ValidationError
