+++ external-import/dragos/tests/tests_dragos/test_domain/test_models/test_observables.py
@@ -11,7 +11,7 @@ from dragos.domain.models.octi import (
     TLPMarking,
     Url,
 )
-from dragos.domain.models.octi.enum import EncryptionAlgorithm, TLPLevel
+from dragos.domain.models.octi.enums import EncryptionAlgorithm, TLPLevel
 from pydantic import ValidationError
